pyproject.toml
[project]
name = "streamlit-mcp-client"
version = "0.1.0"
description = "Streamlit client for MCP Pandas Server with chat interface"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.28.0",
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "openai>=1.3.0",
    "pandas>=2.0.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "asyncio-mqtt>=0.11.0",
    "streamlit-option-menu>=0.3.6",
    "plotly>=5.17.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88